change selectPlugin context name; have consumers manage entity callback reactions
